refactor: delete `AssertionConfiguration` and `KeystoreConfiguration`, update references to use CSL equivalents
